+++ mvlearn/decomposition/multiviewica.py
@@ -29,24 +29,22 @@
 # Modified from source package https://github.com/hugorichard/multiviewica
 
 import numpy as np
-from scipy import linalg
-from sklearn.decomposition import PCA
-from sklearn.utils.validation import check_is_fitted
 
 try:
     from multiviewica import multiviewica
 except ModuleNotFoundError as error:
-    msg = (f"ModuleNotFoundError: {error}. multiviewica dependencies" +
-           "required for this function. Please consult the mvlearn" +
-           "installation instructions at https://github.com/mvlearn/mvlearn" +
-           "to correctly install multiviewica dependencies.")
+    msg = (
+        f"ModuleNotFoundError: {error}. multiviewica dependencies"
+        + "required for this function. Please consult the mvlearn"
+        + "installation instructions at https://github.com/mvlearn/mvlearn"
+        + "to correctly install multiviewica dependencies."
+    )
     raise ModuleNotFoundError(msg)
 
-from .base import BaseDecomposer
-from ..compose import ViewTransformer
+from .base import BaseMultiView
 
 
-class MultiviewICA(BaseDecomposer):
+class MultiviewICA(BaseMultiView):
     r"""
     Multiview ICA for which views share a common source but separate mixing
     matrices.
@@ -179,6 +177,11 @@ class MultiviewICA(BaseDecomposer):
         verbose=False,
         n_jobs=30,
     ):
+        super().__init__(
+            n_components=n_components, multiview_output=multiview_output,
+        )
+
+        self.n_jobs = n_jobs
         self.verbose = verbose
         self.n_components = n_components
         self.noise = noise
@@ -188,26 +191,19 @@ class MultiviewICA(BaseDecomposer):
         self.random_state = random_state
         self.multiview_output = multiview_output
 
-    def fit(self, Xs, y=None):
-        """Fits the model.
-
+    def fit_reduced(self, reduced_X):
+        """
+        Fit the model with reduced data
         Parameters
         ----------
-        Xs: list of np arrays of shape (n_voxels, n_samples)
-            Input data: X[i] is the data of subject i
+        reduced_X: np array of shape (n_views, n_samples, n_components)
 
-        y : ignored
+        Returns
+        --------
+        unmixing: np array of shape (n_views, n_components, n_components)
+            such that reduced_X[i].dot(unmixing[i].T) unmixes the data
+        sources: np array of shape (n_samples, n_components)
         """
-        if self.n_components is not None:
-            self.pcas_ = ViewTransformer(
-                PCA(n_components=self.n_components)
-            )
-
-        if self.n_components is not None:
-            reduced_X = self.pcas_.fit_transform(Xs)
-        else:
-            reduced_X = Xs.copy()
-        reduced_X = np.array(reduced_X)
         _, unmixings_, S = multiviewica(
             np.swapaxes(reduced_X, 1, 2),
             noise=self.noise,
@@ -217,89 +213,14 @@ class MultiviewICA(BaseDecomposer):
             tol=self.tol,
             verbose=self.verbose,
         )
-        mixing_ = np.array([np.linalg.pinv(W) for W in unmixings_])
-        self.components_ = unmixings_
-        self.mixing_ = mixing_
-        if self.n_components is not None:
-            pca_components = []
-            for i, estimator in enumerate(
-                self.pcas_.estimators_
-            ):
-                K = estimator.components_
-                pca_components.append(K)
-            self.pca_components_ = np.array(pca_components)
+        return unmixings_, S.T
 
-        if self.n_components is None:
-            self.individual_components_ = unmixings_
-            self.individual_mixing_ = mixing_
-        else:
-            self.individual_mixing_ = []
-            self.individual_components_ = []
-            sources_pinv = linalg.pinv(S.T)
-            for x in Xs:
-                lstq_solution = np.dot(sources_pinv, x)
-                self.individual_components_.append(
-                    linalg.pinv(lstq_solution).T
-                )
-                self.individual_mixing_.append(lstq_solution.T)
-        return self
-
-    def transform(self, X):
-        r"""
-        Recover the sources from each view (apply unmixing matrix).
-
-        Parameters
-        ----------
-        Xs : list of array-likes or numpy.ndarray
-            - Xs length: n_views
-            - Xs[i] shape: (n_samples, n_features_i)
-            Training data to recover a source and unmixing matrices from.
+    def aggregate(self, X_transformed):
+        """
+        Aggregate transformed data to form a unique output
 
         Returns
         -------
-        Xs_new : numpy.ndarray, shape (n_views, n_samples, n_components)
-            The mixed sources from the single source and per-view unmixings.
+        Source: np array of shape (n_samples, n_components)
         """
-
-        if not hasattr(self, "components_"):
-            raise ValueError("The model has not yet been fitted.")
-
-        if self.n_components is not None:
-            transformed_X = self.pcas_.transform(X)
-        else:
-            transformed_X = X.copy()
-        if self.multiview_output:
-            return np.array(
-                [x.dot(w.T) for w, x in zip(self.components_, transformed_X)]
-            )
-        else:
-            return np.mean(
-                [x.dot(w.T) for w, x in zip(self.components_, transformed_X)],
-                axis=0,
-            )
-
-    def inverse_transform(self, X_transformed):
-        r"""
-        Transforms the sources back to the mixed data for each view
-        (apply mixing matrix).
-
-        Parameters
-        ----------
-        X_transformed : list of array-likes or numpy.ndarray
-            The dataset corresponding to transformed data
-        Returns
-        -------
-        Xs_new : numpy.ndarray, shape (n_views, n_samples, n_components)
-            The mixed sources from the single source and per-view unmixings.
-        """
-        check_is_fitted(self, "components_")
-        if self.multiview_output:
-            S_ = np.mean(X_transformed, axis=0)
-        else:
-            S_ = X_transformed
-        inv_red = [S_.dot(w.T) for w in self.mixing_]
-
-        if self.n_components is not None:
-            return self.pcas_.inverse_transform(inv_red)
-        else:
-            return inv_red
+        return np.mean(X_transformed, axis=0)
